# [FIX] account: _check_off_balance: avoid O(1) and O(n^2).