Main
-

[FIX] account: _check_off_balance: avoid O(n) O(n^2). Let us try making it 0(1)

Explanation
-
There is no need to check for every single journal item to know if they will fail.
By look for one case that actually fails is enough

In case of Journal Entries with all items being off-balance let us say 400 lines
That will lead to 400x400 checks = 160,000 those we are in the case of a O(n^2)
![Screen Shot 2021-03-30 at 15 49 03](https://user-images.githubusercontent.com/7598010/113061637-96a93100-916f-11eb-8f15-8a7ab3100f92.png)





--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr