@hbto 2021-04-07
@william-andre I have come with this new approach.

Profiling
-

I have made the original approach:

```
   1 lines

Timer unit: 1e-06 s

Total time: 0.000236 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         2          9.0      4.5      3.8          for line in self:
  3748         1         37.0     37.0     15.7              if line.account_id.internal_group == 'off_balance':
  3749         1        122.0    122.0     51.7                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3751         1         52.0     52.0     22.0                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753         1         16.0     16.0      6.8                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))







   2 lines

Timer unit: 1e-06 s

Total time: 0.000402 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         3         11.0      3.7      2.7          for line in self:
  3748         2         64.0     32.0     15.9              if line.account_id.internal_group == 'off_balance':
  3749         2        213.0    106.5     53.0                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounds.account_id):ts of the journal entry must be of this type'))ts of the journal entry must be of this type'))
  3751         2         86.0     43.0     21.4                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753         2         28.0     14.0      7.0                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))







   3 lines

Timer unit: 1e-06 s

Total time: 0.000610 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         4         16.0      4.0      2.6          for line in self:
  3748         3         95.0     31.7     15.6              if line.account_id.internal_group == 'off_balance':
  3749         3        329.0    109.7     53.9                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3751         3        126.0     42.0     20.7                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753         3         44.0     14.7      7.2                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))







   4 lines

Timer unit: 1e-06 s

Total time: 0.000881 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         5         21.0      4.2      2.4          for line in self:
  3748         4        135.0     33.8     15.3              if line.account_id.internal_group == 'off_balance':
  3749         4        481.0    120.2     54.6                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3751         4        182.0     45.5     20.7                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753         4         62.0     15.5      7.0                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))









  10 lines

Timer unit: 1e-06 s

Total time: 0.002302 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747        11         46.0      4.2      2.0          for line in self:
  3748        10        338.0     33.8     14.7              if line.account_id.internal_group == 'off_balance':
  3749        10       1316.0    131.6     57.2                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3751        10        448.0     44.8     19.5                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753        10        154.0     15.4      6.7                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))







 100 lines

Timer unit: 1e-06 s

Total time: 0.035789 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747       101        385.0      3.8      1.1          for line in self:
  3748       100       3066.0     30.7      8.6              if line.account_id.internal_group == 'off_balance':
  3749       100      26922.0    269.2     75.2                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3751       100       4018.0     40.2     11.2                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753       100       1398.0     14.0      3.9                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))







1000 lines

Timer unit: 1e-06 s

Total time: 1.87423 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747      1001       3897.0      3.9      0.2          for line in self:
  3748      1000      30378.0     30.4      1.6              if line.account_id.internal_group == 'off_balance':
  3749      1000    1785591.0   1785.6     95.3                  if any(a.internal_group != line.account_id.internal_group for a in line.move_id.line_ids.account_id):
  3750                                                               raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3751      1000      40429.0     40.4      2.2                  if line.tax_ids or line.tax_line_id:
  3752                                                               raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3753      1000      13939.0     13.9      0.7                  if line.reconciled:
  3754                                                               raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))






```

And for the new approach using the recordset:

```
   1 lines

Timer unit: 1e-06 s

Total time: 0.000339 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         1          2.0      2.0      0.6          checked_moves = set()
  3748         2          8.0      4.0      2.4          for line in self:
  3749         1         38.0     38.0     11.2              if line.account_id.internal_group != 'off_balance':
  3750                                                           continue
  3751         1         23.0     23.0      6.8              if line.move_id.id in checked_moves:
  3752                                                           continue
  3753         1         23.0     23.0      6.8              checked_moves.add(line.move_id.id)
  3754         1         84.0     84.0     24.8              if any(a.internal_group != 'off_balance' for a in line.move_id.line_ids.account_id):
  3755                                                           raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3756         1         98.0     98.0     28.9              if any(True for l in line.move_id.line_ids if l.tax_ids or l.tax_line_id):
  3757                                                           raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3758         1         63.0     63.0     18.6              if any(True for l in line.move_id.line_ids if l.reconciled):
  3759                                                           raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))


   2 lines

Timer unit: 1e-06 s

Total time: 0.000430 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         1          2.0      2.0      0.5          checked_moves = set()
  3748         3         13.0      4.3      3.0          for line in self:
  3749         2         65.0     32.5     15.1              if line.account_id.internal_group != 'off_balance':
  3750                                                           continue
  3751         2         42.0     21.0      9.8              if line.move_id.id in checked_moves:
  3752         1          1.0      1.0      0.2                  continue
  3753         1         20.0     20.0      4.7              checked_moves.add(line.move_id.id)
  3754         1         81.0     81.0     18.8              if any(a.internal_group != 'off_balance' for a in line.move_id.line_ids.account_id):
  3755                                                           raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3756         1        132.0    132.0     30.7              if any(True for l in line.move_id.line_ids if l.tax_ids or l.tax_line_id):
  3757                                                           raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3758         1         74.0     74.0     17.2              if any(True for l in line.move_id.line_ids if l.reconciled):
  3759                                                           raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))


   3 lines

Timer unit: 1e-06 s

Total time: 0.000544 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         1          2.0      2.0      0.4          checked_moves = set()
  3748         4         18.0      4.5      3.3          for line in self:
  3749         3         96.0     32.0     17.6              if line.account_id.internal_group != 'off_balance':
  3750                                                           continue
  3751         3         60.0     20.0     11.0              if line.move_id.id in checked_moves:
  3752         2          3.0      1.5      0.6                  continue
  3753         1         21.0     21.0      3.9              checked_moves.add(line.move_id.id)
  3754         1         83.0     83.0     15.3              if any(a.internal_group != 'off_balance' for a in line.move_id.line_ids.account_id):
  3755                                                           raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3756         1        172.0    172.0     31.6              if any(True for l in line.move_id.line_ids if l.tax_ids or l.tax_line_id):
  3757                                                           raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3758         1         89.0     89.0     16.4              if any(True for l in line.move_id.line_ids if l.reconciled):
  3759                                                           raise UserError(_('Lines from "Off-Balance Sheet" accounts cannot be reconciled'))


   4 lines

Timer unit: 1e-06 s

Total time: 0.000675 s
File: /.repo_requirements/odoo/addons/account/models/account_move.py
Function: _check_off_balance at line 3744

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  3744                                               @api.constrains('account_id', 'tax_ids', 'tax_line_id', 'reconciled')
  3745                                               @do_profile()
  3746                                               def _check_off_balance(self):
  3747         1          3.0      3.0      0.4          checked_moves = set()
  3748         5         21.0      4.2      3.1          for line in self:
  3749         4        129.0     32.2     19.1              if line.account_id.internal_group != 'off_balance':
  3750                                                           continue
  3751         4         82.0     20.5     12.1              if line.move_id.id in checked_moves:
  3752         3          5.0      1.7      0.7                  continue
  3753         1         21.0     21.0      3.1              checked_moves.add(line.move_id.id)
  3754         1         85.0     85.0     12.6              if any(a.internal_group != 'off_balance' for a in line.move_id.line_ids.account_id):
  3755                                                           raise UserError(_('If you want to use "Off-Balance Sheet" accounts, all the accounts of the journal entry must be of this type'))
  3756         1        222.0    222.0     32.9              if any(True for l in line.move_id.line_ids if l.tax_ids or l.tax_line_id):
  3757                                                           raise UserError(_('You cannot use taxes on lines with an Off-Balance account'))
  3758         1        107.0    107.0     15.9              if any(True for l in line.move_id.line_ids if l.reconciled):
…